+++ Leomard/Models/Lemmy/Forms/SaveUserSettings.swift
@@ -1,0 +1,14 @@
+//
+//  SaveUserSettings.swift
+//  Leomard
+//
+//  Created by Konrad Figura on 26/07/2023.
+//
+
+import Foundation
+
+struct SaveUserSettings: Codable {
+    let bio: String?
+    let displayName: String?
+}
+