+++ CHANGELOG.md
@@ -14,10 +14,15 @@
 - Cross-posting
 - Hide instances
 - Compact View
+- Added profile editing
+    - You can now change display name and banner in the profile view
+    - You can view blocked communities and persons, as well as unblock them
+    - To access, go into your profile and click the pen icon
 
 ### Changes
 
 - Decreased the indentation of the subcomments
+- Decreased the minimum window height
 
 ### Bug Fixes